% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{kldiv}
\title{Deprecated functions}
\usage{
kldiv(sigma1, beta1, sigma2, beta2, eps = 1e-06)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=kldggd]{kldggd()}} instead of \code{kldiv()}.
}
\keyword{internal}
