% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pochhammer.R
\name{pochhammer}
\alias{pochhammer}
\title{Pochhammer Symbol}
\usage{
pochhammer(x, n)
}
\arguments{
\item{x}{numeric.}

\item{n}{positive integer.}
}
\value{
Numeric value. The value of the Pochhammer symbol.
}
\description{
Computes the Pochhammer symbol.
}
\details{
The Pochhammer symbol is given by:
\deqn{ \displaystyle{ (x)_n = \frac{\Gamma(x+n)}{\Gamma(x)} = x (x+1) ... (x+n-1) } }
}
\examples{
pochhammer(2, 0)
pochhammer(2, 1)
pochhammer(2, 3)

}
\author{
Pierre Santagostini, Nizar Bouhlel
}
